pyproject.toml
[project]
name = "opencontext"
version = "0.1.0"
description = "Local context accumulation engine for AI coding agents"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
oc = "opencontext.cli:main"
